backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.4
motor==3.5.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
openai==1.30.1
chromadb==0.5.5
tenacity==8.2.3
tiktoken==0.7.0
pypdf==4.2.0
numpy==1.26.4
